add concurrency to generate_db_metadata, handle additional columns in db csvs, update prompt
