+++ README.md
@@ -11,11 +11,13 @@ poetry install
 poetry shell
 
 # minidev generate db metadata
-ANTHROPIC_API_KEY=sk-xxx \
-python -m arcwise.generate_db_metadata.main \
+
+AWS_ACCESS_KEY_ID=xxx \
+AWS_SECRET_ACCESS_KEY=xxx \
+AWS_REGION_NAME=us-east-1 \
+python -m arcwise.generate_db_metadata \
   --db-path bird_evaluation/data/dev_databases \
   --tables-json bird_evaluation/data/dev_tables.json \
-  --output-schemas-path bird_evaluation/data/schemas \
   --output-metadata-file bird_evaluation/data/dev_metadata.json
